+++ pex/resolve/resolvers.py
@@ -74,4 +74,3 @@ class InstalledDistribution(object):
 @attr.s(frozen=True)
 class Installed(object):
     installed_distributions = attr.ib()  # type: Tuple[InstalledDistribution, ...]
-    locks = attr.ib(default=())  # type: Tuple[LockedResolve, ...]
