+++ tests/integration/test_locked_resolve.py
@@ -7,54 +7,67 @@ import os
 import pytest
 
 from pex import dist_metadata, resolver, targets
-from pex.resolve.locked_resolve import LockConfiguration, LockStyle
+from pex.fetcher import URLFetcher
+from pex.pip.tool import PackageIndexConfiguration
+from pex.resolve.locked_resolve import LockConfiguration, LockedResolve, LockStyle
+from pex.resolve.lockfile.operations import LockObserver
 from pex.resolve.resolved_requirement import Pin
+from pex.resolve.resolver_configuration import PipConfiguration
 from pex.resolve.testing import normalize_locked_resolve
-from pex.resolver import Downloaded, LocalDistribution
+from pex.resolver import Downloaded, LocalDistribution, WheelBuilder
 from pex.typing import TYPE_CHECKING
 from pex.util import CacheHelper
 
 if TYPE_CHECKING:
-    from typing import Any, Dict, Iterable
-
-    import attr  # vendor:skip
-else:
-    from pex.third_party import attr
-
-
-def normalize_local_dist(local_dist):
-    # type: (LocalDistribution) -> LocalDistribution
-
-    # Each download uses unique temporary dirs as download targets, so paths vary.
-    return attr.evolve(local_dist, path=os.path.basename(local_dist.path))
+    from typing import Any, Dict, Iterable, Tuple
 
 
 def normalize(
-    downloaded,  # type: Downloaded
+    locked_resolves,  # type: Tuple[LockedResolve, ...]
     skip_additional_artifacts=False,  # type: bool
     skip_urls=False,  # type: bool
     skip_verified=False,  # type: bool
 ):
-    # type: (...) -> Downloaded
-    return attr.evolve(
-        downloaded,
-        local_distributions=tuple(
-            sorted(
-                normalize_local_dist(local_dist) for local_dist in downloaded.local_distributions
-            )
-        ),
-        locked_resolves=tuple(
-            sorted(
-                normalize_locked_resolve(
-                    lock,
-                    skip_additional_artifacts=skip_additional_artifacts,
-                    skip_urls=skip_urls,
-                    skip_verified=skip_verified,
-                )
-                for lock in downloaded.locked_resolves
-            )
-        ),
+    # type: (...) -> Tuple[LockedResolve, ...]
+    return tuple(
+        normalize_locked_resolve(
+            lock,
+            skip_additional_artifacts=skip_additional_artifacts,
+            skip_urls=skip_urls,
+            skip_verified=skip_verified,
+        )
+        for lock in locked_resolves
     )
+
+
+def create_lock_observer(lock_configuration):
+    # type: (LockConfiguration) -> LockObserver
+    pip_configuration = PipConfiguration()
+    return LockObserver(
+        lock_configuration=lock_configuration,
+        wheel_builder=WheelBuilder(
+            package_index_configuration=PackageIndexConfiguration.create(
+                resolver_version=pip_configuration.resolver_version,
+                indexes=pip_configuration.repos_configuration.indexes,
+                find_links=pip_configuration.repos_configuration.find_links,
+                network_configuration=pip_configuration.network_configuration,
+            ),
+            prefer_older_binary=pip_configuration.prefer_older_binary,
+            use_pep517=pip_configuration.use_pep517,
+            build_isolation=pip_configuration.build_isolation,
+        ),
+        url_fetcher=URLFetcher(network_configuration=pip_configuration.network_configuration),
+    )
+
+
+def create_lock(
+    lock_configuration,  # type: LockConfiguration
+    **kwargs  # type: Any
+):
+    # type: (...) -> Tuple[Downloaded, Tuple[LockedResolve, ...]]
+    lock_observer = create_lock_observer(lock_configuration)
+    downloaded = resolver.download(observer=lock_observer, **kwargs)
+    return downloaded, lock_observer.lock(downloaded)
 
 
 @pytest.mark.parametrize(
@@ -80,10 +93,9 @@ def test_lock_single_target(
 ):
     # type: (...) -> None
 
-    downloaded = resolver.download(requirements=requirements, lock_configuration=lock_configuration)
-
-    assert 1 == len(downloaded.locked_resolves)
-    lock = downloaded.locked_resolves[0]
+    downloaded, locked_resolves = create_lock(lock_configuration, requirements=requirements)
+    assert 1 == len(locked_resolves)
+    lock = locked_resolves[0]
 
     assert targets.current().platform.tag == lock.platform_tag
 
@@ -120,15 +132,16 @@ def test_lock_single_target(
         os.symlink(
             local_dist.path, os.path.join(find_links_repo, os.path.basename(local_dist.path))
         )
+    _, find_links_locked_resolves = create_lock(
+        lock_configuration,
+        requirements=requirements,
+        indexes=[],
+        find_links=[find_links_repo],
+    )
     assert normalize(
-        downloaded, skip_additional_artifacts=True, skip_urls=True, skip_verified=True
+        locked_resolves, skip_additional_artifacts=True, skip_urls=True, skip_verified=True
     ) == normalize(
-        resolver.download(
-            requirements=requirements,
-            lock_configuration=lock_configuration,
-            indexes=[],
-            find_links=[find_links_repo],
-        ),
+        find_links_locked_resolves,
         skip_additional_artifacts=True,
         skip_urls=True,
         skip_verified=True,
@@ -141,9 +154,8 @@ def test_lock_single_target(
     lock_file = os.path.join(str(tmpdir), "requirements.txt")
     with open(lock_file, "w") as fp:
         lock.emit_requirements(fp)
-    assert normalize(downloaded) == normalize(
-        resolver.download(requirement_files=[lock_file], lock_configuration=lock_configuration),
-    ), (
+    _, export_locked_resolves = create_lock(lock_configuration, requirement_files=[lock_file])
+    assert normalize(locked_resolves) == normalize(export_locked_resolves), (
         "Expected the download used to create a lock to be reproduced by a download using the "
         "requirements generated from the lock."
     )
